Rebrand from template to waambokt and add some env configuration and docker deployment files
