--- .github/boring-cyborg.yml
@@ -1,2 +1,0 @@
-# Boring Cyborg: https://probot.github.io/apps/boring-cyborg/
-_extends: ".github"
