--- .github/config.yml
@@ -1,2 +1,0 @@
-# to-do: https://probot.github.io/apps/todo/
-_extends: ".github"
