--- .github/settings.yml
@@ -1,2 +1,0 @@
-# Settings: https://probot.github.io/apps/settings/
-_extends: ".github"
