+++ .github/workflows/root.yml
@@ -3,7 +3,7 @@ name: Build & Publish
 on:
   push:
     branches:
-      - root
+      - main
       - develop
 
 jobs:
