+++ .gitignore
@@ -1,5 +1,5 @@
 # Created by .ignore support plugin (hsz.mobi)
-### Kotlin template
+### Kotlin waambokt
 # Compiled class file
 *.class
 
@@ -24,7 +24,7 @@
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
 
-### JetBrains template
+### JetBrains waambokt
 # Covers JetBrains IDEs: IntelliJ, RubyMine, PhpStorm, AppCode, PyCharm, CLion, Android Studio, WebStorm and Rider
 # Reference: https://intellij-support.jetbrains.com/hc/en-us/articles/206544839
