+++ .idea/jarRepositories.xml
@@ -46,5 +46,10 @@
       <option name="name" value="Google" />
       <option name="url" value="https://dl.google.com/dl/android/maven2/" />
     </remote-repository>
+    <remote-repository>
+      <option name="id" value="Sonatype Snapshots" />
+      <option name="name" value="Sonatype Snapshots" />
+      <option name="url" value="https://s01.oss.sonatype.org/content/repositories/snapshots" />
+    </remote-repository>
   </component>
 </project>
