+++ .idea/misc.xml
@@ -8,4 +8,7 @@
   <component name="ProjectRootManager" version="2" languageLevel="JDK_11" default="true" project-jdk-name="11" project-jdk-type="JavaSDK">
     <output url="file://$PROJECT_DIR$/out" />
   </component>
+  <component name="UnicodeBrowser">
+    <option name="fontName" value="JetBrains Mono" />
+  </component>
 </project>
