--- src/main/kotlin/template/App.kt
@@ -1,40 +1,0 @@
-/*
- * This Kotlin source file was generated by the Gradle 'init' task.
- */
-package template
-
-import com.kotlindiscord.kord.extensions.ExtensibleBot
-import com.kotlindiscord.kord.extensions.utils.env
-import dev.kord.common.entity.Snowflake
-import template.extensions.TestExtension
-
-val TEST_SERVER_ID = Snowflake(
-    env("TEST_SERVER").toLong()  // Get the test server ID from the env vars or a .env file
-)
-
-private val TOKEN = env("TOKEN")   // Get the bot' token from the env vars or a .env file
-
-suspend fun main() {
-    val bot = ExtensibleBot(TOKEN) {
-        chatCommands {
-            defaultPrefix = "?"
-            enabled = true
-
-            prefix { default ->
-                if (guildId == TEST_SERVER_ID) {
-                    // For the test server, we use ! as the command prefix
-                    "!"
-                } else {
-                    // For other servers, we use the configured default prefix
-                    default
-                }
-            }
-        }
-
-        extensions {
-            add(::TestExtension)
-        }
-    }
-
-    bot.start()
-}
